Implemented Changes suggested by review
